+++ crates/it-lilo/src/traits/allocatable.rs
@@ -19,7 +19,7 @@ use thiserror::Error as ThisError;
 pub const DEFAULT_MEMORY_INDEX: usize = 0;
 
 pub trait Allocatable {
-    fn allocate(&self, size: u32, type_tag: u32) -> Result<usize, AllocatableError>;
+    fn allocate(&self, size: u32, type_tag: u32) -> Result<u32, AllocatableError>;
 }
 
 #[derive(Debug, ThisError)]
