+++ crates/it-memory-traits/src/errors.rs
@@ -20,8 +20,8 @@ use thiserror::Error as ThisError;
 pub enum MemoryAccessError {
     #[error("Out-of-bound Wasm memory access: offset {offset}, size {size}, while memory_size {memory_size}")]
     OutOfBounds {
-        offset: usize,
-        size: usize,
-        memory_size: usize,
+        offset: u32,
+        size: u32,
+        memory_size: u32,
     },
 }
